puzzle/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
psycopg2-binary
sqlalchemy
python-dotenv
httpx
loguru 